Issue 50041 - Add the rest UI Plugin tabs - Part 2
Description: Add UI plugin tabs for autoMembership, DNA, managedEntries,
passthroughAuthentication, usn.
Add Shared Config Entry to referentialIntegrity plugin.
Add Plugin Precedence field to the basic plugin configuration.
Fix CLI tools according to the UI changes.

https://pagure.io/389-ds-base/issue/50041

Reviewed by: mreynolds (Thanks!)

+++ src/lib389/lib389/cli_conf/plugins/accountpolicy.py
@@ -32,6 +32,8 @@ def accountpolicy_edit(inst, basedn, log, args):
 def accountpolicy_add_config(inst, basedn, log, args):
     log = log.getChild('accountpolicy_add_config')
     targetdn = args.DN
+    if not ldap.dn.is_dn(targetdn):
+        raise ValueError("Specified DN is not a valid DN")
     config = generic_object_add(AccountPolicyConfig, inst, log, args, arg_to_attr_config, dn=targetdn)
     plugin = AccountPolicyPlugin(inst)
     plugin.replace('nsslapd_pluginConfigArea', config.dn)
@@ -42,6 +44,8 @@ def accountpolicy_add_config(inst, basedn, log, args):
 def accountpolicy_edit_config(inst, basedn, log, args):
     log = log.getChild('accountpolicy_edit_config')
     targetdn = args.DN
+    if not ldap.dn.is_dn(targetdn):
+        raise ValueError("Specified DN is not a valid DN")
     config = AccountPolicyConfig(inst, targetdn)
     generic_object_edit(config, log, args, arg_to_attr_config)
 
@@ -49,6 +53,8 @@ def accountpolicy_edit_config(inst, basedn, log, args):
 def accountpolicy_show_config(inst, basedn, log, args):
     log = log.getChild('accountpolicy_show_config')
     targetdn = args.DN
+    if not ldap.dn.is_dn(targetdn):
+        raise ValueError("Specified DN is not a valid DN")
     config = AccountPolicyConfig(inst, targetdn)
 
     if not config.exists():
@@ -64,6 +70,8 @@ def accountpolicy_show_config(inst, basedn, log, args):
 def accountpolicy_del_config(inst, basedn, log, args):
     log = log.getChild('accountpolicy_del_config')
     targetdn = args.DN
+    if not ldap.dn.is_dn(targetdn):
+        raise ValueError("Specified DN is not a valid DN")
     config = AccountPolicyConfig(inst, targetdn)
     config.delete()
     log.info("Successfully deleted the %s", targetdn)
